==> docs/step6_check_performance.ipynb <==
import pandas as pd

arima_cv = pd.read_csv("./models/CV_ARIMA.csv").drop('Unnamed: 0', 1)
lr_cv = pd.read_csv("./models/CV_LR.csv").drop('Unnamed: 0', 1)
trees_cv = pd.read_csv("./models/CV_Trees.csv").drop('Unnamed: 0', 1)

total = pd.concat([arima_cv, lr_cv, trees_cv], 1)

pd.DataFrame({'avg': total.mean(), 'std': total.std()}).sort_values(by='avg', ascending=False)

Давайте загрузим тестовый датасет и посмотрим прогнозы на нем.

from sklearn.metrics import mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.preprocessing import StandardScaler

train = pd.read_csv("train_prepared.csv", index_col=[0], parse_dates=[0])
test = pd.read_csv("test_prepared.csv", index_col=[0], parse_dates=[0])

sc = StandardScaler()
X_train = train.drop('Passengers', 1).values
y_train = train['Passengers'].values
scaler = sc.fit(X_train)
X_train = sc.transform(X_train)

#test prepare
X_test = test.drop('Passengers', 1).values
y_test = test['Passengers'].values
X_test = sc.transform(X_test)

model = Lasso()
model.fit(X_train, y_train)
preds = model.predict(X_test)
r2_score(y_test, preds)

model.coef_

features_imps = pd.DataFrame({'feature': [col for col in train.drop('Passengers', 1).columns], 
              'coef_': model.coef_})
features_imps[features_imps['coef_']==0]

features_imps[features_imps['coef_']!=0]

Попробуем с другими моделями

model = Ridge()
model.fit(X_train, y_train)
preds = model.predict(X_test)
r2_score(y_test, preds)

model = LinearRegression()
model.fit(X_train, y_train)
preds = model.predict(X_test)
r2_score(y_test, preds)

Лучшей моделью выбираем пока LASSO

model = Lasso()
model.fit(X_train, y_train)

from statsmodels.api import OLS
model = OLS(y_train, X_train).fit_regularized(alpha=0.8, L1_wt=1.0)

regularized_regression_parameters = model.params
print(regularized_regression_parameters)

features_imps = pd.DataFrame({'feature': [col for col in train.drop('Passengers', 1).columns], 
              'coef_': regularized_regression_parameters})
features_imps

selected_features = features_imps[features_imps['coef_']!=0]['feature'].values

sc = StandardScaler()
X_train = train.drop('Passengers', 1)[selected_features].values
y_train = train['Passengers'].values
scaler = sc.fit(X_train)
X_train = sc.transform(X_train)

#test prepare
X_test = test.drop('Passengers', 1)[selected_features].values
y_test = test['Passengers'].values
X_test = sc.transform(X_test)

model = Lasso()
model.fit(X_train, y_train)
preds = model.predict(X_test)
r2_score(y_test, preds)
